Fix compile MSVC

+++ apps/openmw_test_suite/mwworld/test_store.cpp
@@ -315,8 +315,8 @@ static void testRecNameIntCount(const MWWorld::Store<T>& store, const MWWorld::E
     {
         const unsigned int recordIdCount
             = std::apply([](auto&&... x) { return (hasSameRecordId(x, T::sRecordId) + ...); }, stores);
-        ASSERT_EQ(recordIdCount, 1) << "The same RecNameInt is used twice ESM::REC_"
-                                    << MWWorld::getRecNameString(T::sRecordId).toStringView();
+        ASSERT_EQ(recordIdCount, static_cast<unsigned int>(1))
+            << "The same RecNameInt is used twice ESM::REC_" << MWWorld::getRecNameString(T::sRecordId).toStringView();
     }
 }
